image.sh: riscv64 EFI bootloader dir and uncompressed vmlinux support

+++ build/scripts/lib/image.sh
@@ -454,7 +454,15 @@ SSHEOF
         -b 1048576 -comp xz -Xdict-size 100% -xattrs
 
     log_step "Copying kernel and initramfs..."
-    cp "$_chroot_dir/boot/vmlinuz-$_imagekernelversion" "$_basedir/image_tree/image/vmlinuz"
+    # riscv64's linux-image ships an uncompressed vmlinux-<ver> (no vmlinuz-);
+    # the /vmlinuz symlink in the chroot root points at whichever exists.
+    if [ ! -f "$_chroot_dir/boot/vmlinuz-$_imagekernelversion" ] \
+       && [ -f "$_chroot_dir/boot/vmlinux-$_imagekernelversion" ]; then
+        log_step "riscv64: using uncompressed vmlinux-$_imagekernelversion"
+        cp "$_chroot_dir/boot/vmlinux-$_imagekernelversion" "$_basedir/image_tree/image/vmlinuz"
+    else
+        cp "$_chroot_dir/boot/vmlinuz-$_imagekernelversion" "$_basedir/image_tree/image/vmlinuz"
+    fi
     cp "$_chroot_dir/boot/initrd.img-$_imagekernelversion" "$_basedir/image_tree/image/initrd"
 
     log_step "Creating GRUB menu..."
@@ -524,7 +532,7 @@ EOF
         sudo mkdir -p "$_basedir/image_tree/chroot/scratch"
         sudo cp "$_basedir/image_tree/scratch/grub.cfg" "$_basedir/image_tree/chroot/scratch/"
         sudo chroot "$_basedir/image_tree/chroot" /usr/bin/grub-mkstandalone \
-            --directory=/usr/lib/grub/arm64-efi \
+            --directory="/usr/lib/grub/$_efi_target" \
             --format="$_efi_target" \
             --output="/scratch/$_efi_name" \
             --locales="" \
